+++ packages/xchain-thornode/src/config.ts
@@ -1,0 +1,3 @@
+export const THORNODE_API_9R_URL = 'https://thornode.ninerealms.com/';
+
+